chore: clean up naming inconsistencies and duplicate modules
Remove 3 duplicate module directories left over from an incomplete
reorganization (context/, knowledge/, name_policies/), consolidate
their tests into the canonical analysis/ and config/ locations,
rename MemoryMappedScanner/ParallelShardScanner to *Handler to
avoid confusion with BaseScanner subclasses, fix _registry leak
in __all__, and add 3 missing scanner entries to lazy-load map.
