--- modelaudit/context/__init__.py
@@ -1,5 +1,0 @@
-"""Context management for unified ML analysis."""
-
-from .unified_context import LayerPattern, ModelArchitecture, TensorInfo, UnifiedMLContext
-
-__all__ = ["LayerPattern", "ModelArchitecture", "TensorInfo", "UnifiedMLContext"]
