--- modelaudit/knowledge/__init__.py
@@ -1,5 +1,0 @@
-"""Framework-specific knowledge base."""
-
-from .framework_patterns import FrameworkKnowledgeBase, FrameworkPattern, FrameworkType
-
-__all__ = ["FrameworkKnowledgeBase", "FrameworkPattern", "FrameworkType"]
